@@ -31,30 +31,31 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "~10.1.0",
-    "@angular/common": "~10.1.0",
-    "@angular/compiler": "~10.1.0",
-    "@angular/core": "~10.1.0",
-    "@angular/forms": "~10.1.0",
-    "@angular/localize": "~10.1.0",
-    "@angular/platform-browser": "~10.1.0",
-    "@angular/platform-browser-dynamic": "~10.1.0",
-    "@angular/platform-server": "~10.1.0",
-    "@angular/router": "~10.1.0",
-    "@angular/service-worker": "~10.1.0",
+    "@angular/animations": "~10.2.4",
+    "@angular/common": "~10.2.4",
+    "@angular/compiler": "~10.2.4",
+    "@angular/core": "~10.2.4",
+    "@angular/forms": "~10.2.4",
+    "@angular/localize": "~10.2.4",
+    "@angular/platform-browser": "~10.2.4",
+    "@angular/platform-browser-dynamic": "~10.2.4",
+    "@angular/platform-server": "~10.2.4",
+    "@angular/router": "~10.2.4",
+    "@angular/service-worker": "~10.2.4",
     "@compodoc/compodoc": "^1.1.10",
     "@ng-bootstrap/ng-bootstrap": "^7.0.0",
-    "@ng-select/ng-select": "^5.0.0",
+    "@ng-select/ng-select": "^5.0.9",
     "@ngrx/effects": "~10.0.0",
     "@ngrx/router-store": "~10.0.0",
     "@ngrx/store": "~10.0.0",
     "@nguniversal/express-engine": "^10.1.0",
-    "@spartacus/assets": "3.0.0",
-    "@spartacus/core": "3.0.0",
-    "@spartacus/organization": "3.0.0",
-    "@spartacus/storefinder": "3.0.0",
-    "@spartacus/storefront": "3.0.0",
-    "@spartacus/styles": "3.0.0",
+    "@spartacus/assets": "3.4.0",
+    "@spartacus/core": "3.4.0",
+    "@spartacus/organization": "3.4.0",
+    "@spartacus/storefinder": "3.4.0",
+    "@spartacus/storefront": "3.4.0",
+    "@spartacus/user": "3.4.0",
+    "@spartacus/styles": "3.4.0",
     "@types/googlemaps": "^3.37.5",
     "angular-oauth2-oidc": "^10.0.3",
     "blob-util": "^2.0.2",
@@ -77,12 +78,12 @@
     "@angular-builders/dev-server": "^7.3.1",
     "@angular-devkit/build-angular": "~0.1001.0",
     "@angular-devkit/build-ng-packagr": "~0.1001.0",
-    "@angular/cli": "~10.1.0",
-    "@angular/compiler-cli": "~10.1.0",
-    "@angular/language-service": "~10.1.0",
+    "@angular/cli": "~10.2.1",
+    "@angular/compiler-cli": "~10.2.4",
+    "@angular/language-service": "~10.2.4",
     "@ngrx/store-devtools": "~10.0.0",
     "@nguniversal/builders": "^10.1.0",
-    "@spartacus/schematics": "3.0.0",
+    "@spartacus/schematics": "3.4.0",
     "@types/express": "^4.17.0",
     "@types/i18next": "^12.1.0",
     "@types/jasmine": "~3.5.0",
