@@ -15,19 +15,20 @@
     "storefront"
   ],
   "peerDependencies": {
-    "@angular/common": "~10.1.0",
-    "@angular/compiler": "~10.1.0",
-    "@angular/core": "~10.1.0",
-    "@angular/forms": "~10.1.0",
-    "@angular/localize": "~10.1.0",
-    "@angular/platform-browser": "~10.1.0",
-    "@angular/router": "~10.1.0",
-    "@angular/service-worker": "~10.1.0",
+    "@angular/common": "~10.2.4",
+    "@angular/compiler": "~10.2.4",
+    "@angular/core": "~10.2.4",
+    "@angular/forms": "~10.2.4",
+    "@angular/localize": "~10.2.4",
+    "@angular/platform-browser": "~110.2.4",
+    "@angular/router": "~10.2.4",
+    "@angular/service-worker": "~10.2.4",
     "@ng-bootstrap/ng-bootstrap": "^6.0.0",
     "@ng-select/ng-select": "^5.0.0",
     "@ngrx/effects": "~10.0.0",
     "@ngrx/store": "~10.0.0",
-    "@spartacus/core": "3.0.0",
+    "@spartacus/core": "3.4.0",
+    "@spartacus/user": "3.4.0",
     "bootstrap": "4.5.0",
     "echarts": "^5.0.2",
     "ngx-echarts": "^6.0.1",
