@@ -1,8 +1,9 @@
 import { TestBed } from '@angular/core/testing';
 import { StoreModule } from '@ngrx/store';
-import { I18nTestingModule, UserService } from '@spartacus/core';
+import { I18nTestingModule } from '@spartacus/core';
 import { of } from 'rxjs';
 import { UserPrefillResolver } from './user-prefill-resolver';
+import { UserAccountFacade } from '@spartacus/user/account/root';
 
 const firstName = 'Donna';
 const lastName = 'Moore';
@@ -12,7 +13,7 @@ const mockUser = {
   lastName: lastName,
 };
 
-class MockUserService {
+class MockUserAccountFacade {
   get() {
     return of(mockUser);
   }
@@ -22,16 +23,18 @@ const mockFieldPath = 'firstName';
 
 describe('UserPrefilResolver', () => {
   let userPrefilResolver: UserPrefillResolver;
-  let userService: UserService;
+  let mockedUserAccountFacade: UserAccountFacade;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [I18nTestingModule, StoreModule.forRoot({})],
-      providers: [{ provide: UserService, useClass: MockUserService }],
+      providers: [
+        { provide: UserAccountFacade, useClass: MockUserAccountFacade },
+      ],
     });
 
     userPrefilResolver = TestBed.inject(UserPrefillResolver);
-    userService = TestBed.inject(UserService);
+    mockedUserAccountFacade = TestBed.inject(UserAccountFacade);
   });
 
   it('should inject user resolver', () => {
