@@ -1,57 +1,61 @@
 export function searchNoResults() {
-  cy.get('cx-searchbox')
+  cy.get('cx-fs-searchbox')
     .should('be.visible')
     .within(() => {
-      cy.get('input').type('notexist{enter}').wait(1500);
+      cy.get('input').type('notexist{enter}');
     });
   cy.get('cx-breadcrumb').should('contain', '0 results for "notexist"');
 }
 
 export function searchInsuranceProducts() {
-  cy.get('cx-searchbox')
+  cy.get('cx-fs-searchbox')
     .should('be.visible')
     .within(() => {
-      cy.get('input').clear().type('insurance{enter}').wait(1500);
+      cy.get('input').clear().type('insurance{enter}');
     });
   cy.get('cx-breadcrumb').should('contain', '25 results for "insurance"');
 }
 
 export function searchBankingProducts() {
-  cy.get('cx-searchbox')
+  cy.get('cx-fs-searchbox')
     .should('be.visible')
     .within(() => {
-      cy.get('input').clear().type('account{enter}').wait(1500);
+      cy.get('input').clear().type('account{enter}');
     });
   cy.get('cx-breadcrumb').should('contain', '3 results for "account"');
 }
 
-export function searchSavingsProducts() {
-  cy.get('cx-searchbox')
+export function searchTravelProducts() {
+  cy.get('cx-fs-searchbox')
     .should('be.visible')
     .within(() => {
-      cy.get('input').clear().type('savings{enter}').wait(1500);
+      cy.get('input').clear().type('travel{enter}');
     });
-  cy.get('cx-breadcrumb').should('contain', '3 results for "savings"');
+  cy.get('cx-breadcrumb').should('contain', '9 results for "travel"');
+  cy.reload();
+  cy.get('.title.background-color-primary')
+    .eq(0)
+    .should('contain', ' Backpackers - Gold Plan ');
 }
 
 export function seachResultsButtons() {
   cy.get('cx-fs-product-list')
     .should('be.visible')
     .within(() => {
-      cy.get('cx-media').should('have.length', 3);
-      cy.get('button.primary-button')
+      cy.get('cx-media').should('have.length', 9);
+      cy.get('button.btn-action')
         .should('contain', 'More Info')
-        .should('have.length', 3);
-      cy.get('button.secondary-button')
+        .should('have.length', 9);
+      cy.get('button.primary-button')
         .should('contain', 'Get a Quote')
-        .should('have.length', 3);
+        .should('have.length', 9);
     });
 }
 
 export function clickMoreInfoButton() {
-  cy.get('button.primary-button').eq(0).click({ force: true }).wait(1500);
+  cy.get('button.btn-action').eq(0).click({ force: true });
 }
 
 export function clickGetAQuoteButton() {
-  cy.get('button.secondary-button').eq(0).click({ force: true }).wait(1500);
+  cy.get('button.primary-button').eq(0).click({ force: true });
 }
