@@ -1,4 +1,5 @@
 import * as productCategory from '../../../helpers/productCategoryPage';
+import { insuranceButtons, bankingButtons } from '../../../helpers/constants';
 
 context('ProductCategoryPage', () => {
   before(() => {
@@ -12,17 +13,11 @@ context('ProductCategoryPage', () => {
       dropdownItem: 'Homeowners',
     });
     productCategory.checkComponents();
-    productCategory.checkQuoteButtons();
-    cy.get('.product-feature-wrapper').should('have.length', 2);
-    cy.get('.section-header-heading').should(
-      'contain.text',
-      'Homeowners Monthly'
-    );
-    cy.get('.section-header-heading').should(
-      'contain.text',
+    productCategory.checkCategoryBannerButtons(insuranceButtons);
+    productCategory.checkCategoryProductHeadings(
+      'Homeowners Monthly',
       'Homeowners Annually'
     );
-    cy.get('.item-details').should('have.length', 2);
   });
 
   it('should check renters product category page', () => {
@@ -31,14 +26,11 @@ context('ProductCategoryPage', () => {
       dropdownItem: 'Renters',
     });
     productCategory.checkComponents();
-    productCategory.checkQuoteButtons();
-    cy.get('.product-feature-wrapper').should('have.length', 2);
-    cy.get('.section-header-heading').should('contain.text', 'Renters Monthly');
-    cy.get('.section-header-heading').should(
-      'contain.text',
+    productCategory.checkCategoryBannerButtons(insuranceButtons);
+    productCategory.checkCategoryProductHeadings(
+      'Renters Monthly',
       'Renters Annually'
     );
-    cy.get('.item-details').should('have.length', 2);
   });
 
   it('should check Auto product category page', () => {
@@ -47,12 +39,12 @@ context('ProductCategoryPage', () => {
       dropdownItem: 'Auto',
     });
     productCategory.checkComponents();
-    productCategory.checkQuoteButtons();
-    cy.get('.product-feature-wrapper').should('have.length', 3);
-    cy.get('.section-header-heading').should('contain.text', 'Auto Bronze');
-    cy.get('.section-header-heading').should('contain.text', 'Auto Silver');
-    cy.get('.section-header-heading').should('contain.text', 'Auto Gold');
-    cy.get('.item-details').should('have.length', 3);
+    productCategory.checkCategoryBannerButtons(insuranceButtons);
+    productCategory.checkCategoryProductHeadings(
+      'Auto Bronze',
+      'Auto Silver',
+      'Auto Gold'
+    );
   });
 
   it('should check Life product category page', () => {
@@ -61,11 +53,8 @@ context('ProductCategoryPage', () => {
       dropdownItem: 'Life',
     });
     productCategory.checkComponents();
-    productCategory.checkQuoteButtons();
-    cy.get('.product-feature-wrapper').should('have.length', 2);
-    cy.get('.section-header-heading').should('contain.text', 'Life Basic');
-    cy.get('.section-header-heading').should('contain.text', 'Life Premium');
-    cy.get('.item-details').should('have.length', 2);
+    productCategory.checkCategoryBannerButtons(insuranceButtons);
+    productCategory.checkCategoryProductHeadings('Life Basic', 'Life Premium');
   });
 
   it('should check Travel product category page', () => {
@@ -74,12 +63,12 @@ context('ProductCategoryPage', () => {
       dropdownItem: 'Travel',
     });
     productCategory.checkComponents();
-    productCategory.checkQuoteButtons();
-    cy.get('.product-feature-wrapper').should('have.length', 3);
-    cy.get('.section-header-heading').should('contain.text', 'Single Budget');
-    cy.get('.section-header-heading').should('contain.text', 'Single Silver');
-    cy.get('.section-header-heading').should('contain.text', 'Single Gold');
-    cy.get('.item-details').should('have.length', 3);
+    productCategory.checkCategoryBannerButtons(insuranceButtons);
+    productCategory.checkCategoryProductHeadings(
+      'Single Budget',
+      'Single Silver',
+      'Single Gold'
+    );
   });
 
   it('should check Event product category page', () => {
@@ -88,18 +77,12 @@ context('ProductCategoryPage', () => {
       dropdownItem: 'Event',
     });
     productCategory.checkComponents();
-    productCategory.checkQuoteButtons();
-    cy.get('.product-feature-wrapper').should('have.length', 3);
-    cy.get('.section-header-heading').should('contain.text', 'Event Two Stars');
-    cy.get('.section-header-heading').should(
-      'contain.text',
-      'Event Three Stars'
-    );
-    cy.get('.section-header-heading').should(
-      'contain.text',
+    productCategory.checkCategoryBannerButtons(insuranceButtons);
+    productCategory.checkCategoryProductHeadings(
+      'Event Two Stars',
+      'Event Three Stars',
       'Event Four Stars'
     );
-    cy.get('.item-details').should('have.length', 3);
   });
 
   it('should check Savings product category page', () => {
@@ -108,8 +91,12 @@ context('ProductCategoryPage', () => {
       dropdownItem: 'Savings',
     });
     productCategory.checkComponents();
-    productCategory.checkQuoteButtons();
-    productCategory.checksSavingsCategoryPage();
+    productCategory.checkCategoryBannerButtons(insuranceButtons);
+    productCategory.checkCategoryProductHeadings(
+      'Savings Safe And Steady',
+      'Savings Balanced Deal',
+      'Savings Flexi-Max'
+    );
   });
 
   it('should check Current Account product category page', () => {
@@ -118,12 +105,12 @@ context('ProductCategoryPage', () => {
       dropdownItem: 'Current Account',
     });
     productCategory.checkComponents();
-    productCategory.checkApplicationButtons();
-    cy.get('.product-feature-wrapper').should('have.length', 3);
-    cy.get('.section-header-heading').should('contain.text', 'Basic Account');
-    cy.get('.section-header-heading').should('contain.text', 'Family Account');
-    cy.get('.section-header-heading').should('contain.text', 'Premium Account');
-    cy.get('.item-details').should('have.length', 3);
+    productCategory.checkCategoryBannerButtons(bankingButtons);
+    productCategory.checkCategoryProductHeadings(
+      'Basic Account',
+      'Family Account',
+      'Premium Account'
+    );
   });
 
   it('should check Credit Card product category page', () => {
@@ -132,12 +119,12 @@ context('ProductCategoryPage', () => {
       dropdownItem: 'Credit Card',
     });
     productCategory.checkComponents();
-    productCategory.checkApplicationButtons();
-    cy.get('.product-feature-wrapper').should('have.length', 3);
-    cy.get('.section-header-heading').should('contain.text', 'Basic Card');
-    cy.get('.section-header-heading').should('contain.text', 'Premium Card');
-    cy.get('.section-header-heading').should('contain.text', 'Exclusive Card');
-    cy.get('.item-details').should('have.length', 3);
+    productCategory.checkCategoryBannerButtons(bankingButtons);
+    productCategory.checkCategoryProductHeadings(
+      'Basic Card',
+      'Premium Card',
+      'Exclusive Card'
+    );
   });
 
   it('should check Loan product category page', () => {
@@ -146,10 +133,8 @@ context('ProductCategoryPage', () => {
       dropdownItem: 'Loan',
     });
     productCategory.checkComponents();
-    productCategory.checkApplicationButtons();
-    cy.get('.product-feature-wrapper').should('have.length', 1);
-    cy.get('.section-header-heading').should('contain.text', 'Personal Loan');
-    cy.get('.item-details').should('have.length', 1);
+    productCategory.checkCategoryBannerButtons(bankingButtons);
+    productCategory.checkCategoryProductHeadings('Personal Loan');
   });
 
   it('should check Fixed Term Deposit product category page', () => {
@@ -158,12 +143,7 @@ context('ProductCategoryPage', () => {
       dropdownItem: 'Fixed Term Deposit',
     });
     productCategory.checkComponents();
-    productCategory.checkApplicationButtons();
-    cy.get('.product-feature-wrapper').should('have.length', 1);
-    cy.get('.section-header-heading').should(
-      'contain.text',
-      'Fixed Term Deposit'
-    );
-    cy.get('.item-details').should('have.length', 1);
+    productCategory.checkCategoryBannerButtons(bankingButtons);
+    productCategory.checkCategoryProductHeadings('Fixed Term Deposit');
   });
 });
