@@ -1,11 +1,11 @@
 import * as productCategory from '../../../helpers/productCategoryPage';
 import * as search from '../../../helpers/search';
+import { insuranceButtons } from '../../../helpers/constants';
 
 context('Search', () => {
   before(() => {
     cy.visit('/');
   });
-
   //TODO: Bug FSA-5142
   it('Should search world that will lead to no results page', () => {
     search.searchNoResults();
@@ -20,13 +20,18 @@ context('Search', () => {
   });
 
   it('Should check buttons on search results', () => {
-    search.searchSavingsProducts();
+    search.searchTravelProducts();
     search.seachResultsButtons();
     search.clickMoreInfoButton();
     productCategory.checkComponents();
-    productCategory.checkQuoteButtons();
-    productCategory.checksSavingsCategoryPage();
+    productCategory.checkCategoryBannerButtons(insuranceButtons);
+    productCategory.checkCategoryProductHeadings(
+      'Travel Single Silver',
+      'Travel Single Budget',
+      'Travel Single Gold'
+    );
+    cy.go('back');
     search.clickGetAQuoteButton();
-    cy.get('h2').should('contain', 'Your Savings Insurance');
+    cy.get('h2').should('contain', 'Your Travel Insurance');
   });
 });
