@@ -4,13 +4,13 @@ import {
   TranslationService,
   User,
   UserAddressService,
-  UserService,
 } from '@spartacus/core';
 import {
   AddressBookComponent,
   AddressBookComponentService,
 } from '@spartacus/storefront';
 import { Observable } from 'rxjs';
+import { UserAccountFacade } from '@spartacus/user/account/root';
 
 @Component({
   templateUrl: './address-info.component.html',
@@ -24,13 +24,13 @@ export class FSAddressInfoComponent extends AddressBookComponent
     protected translation: TranslationService,
     protected userAddressService: UserAddressService,
     protected checkoutDeliveryService: CheckoutDeliveryService,
-    protected userService: UserService
+    protected userAccountFacade: UserAccountFacade
   ) {
     super(service, translation, userAddressService, checkoutDeliveryService);
   }
 
   ngOnInit() {
     super.ngOnInit();
-    this.user$ = this.userService.get();
+    this.user$ = this.userAccountFacade.get();
   }
 }
