@@ -1,4 +1,4 @@
 export * from './policies/policies.component';
 export * from './policy-details/policy-details.component';
-export * from './policies-chart/policies-chart.component';
 export * from './policy.module';
+export * from './policies-chart/policies-chart.module';
