@@ -17,7 +17,6 @@ import {
   RoutingConfig,
 } from '@spartacus/core';
 import { CmsPageGuard } from '@spartacus/storefront';
-
 import { PoliciesComponent } from './policies/policies.component';
 import { PolicyService } from '../../../core/my-account/facade/policy.service';
 import { PolicyDetailsComponent } from './policy-details/policy-details.component';
@@ -25,11 +24,7 @@ import { AccordionModule } from '../../../shared/accordion/accordion.module';
 import { ChangeRequestService } from './../../../core/change-request/facade/change-request.service';
 import { ChangeRequestStoreModule } from './../../../core/change-request/store/change-request-store.module';
 import { DocumentsTableModule } from '../documents/documents-table/documents-table.module';
-import { PoliciesChartComponent } from './policies-chart/policies-chart.component';
-import { NgxEchartsModule } from 'ngx-echarts';
-import * as echarts from 'echarts';
-import { defaultChartOptionsConfig } from '../../../core/chart-config/default-chart-options-config';
-import { FormsModule } from '@angular/forms';
+import { PoliciesChartModule } from './policies-chart/policies-chart.module';
 
 const routes: Routes = [
   {
@@ -55,7 +50,6 @@ const routes: Routes = [
 @NgModule({
   imports: [
     CommonModule,
-    FormsModule,
     RouterModule,
     I18nModule,
     NgSelectModule,
@@ -64,16 +58,10 @@ const routes: Routes = [
     ChangeRequestStoreModule,
     MediaModule,
     DocumentsTableModule,
-    NgxEchartsModule.forRoot({
-      echarts,
-    }),
+    PoliciesChartModule,
     RouterModule.forChild(routes),
-    ConfigModule.withConfig(defaultChartOptionsConfig),
     ConfigModule.withConfig(<CmsConfig | RoutesConfig | RoutingConfig>{
       cmsComponents: {
-        AccountMyPoliciesChartFlex: {
-          component: PoliciesChartComponent,
-        },
         AccountMyPoliciesFlex: {
           component: PoliciesComponent,
         },
@@ -83,17 +71,8 @@ const routes: Routes = [
       },
     }),
   ],
-  declarations: [
-    PoliciesComponent,
-    PolicyDetailsComponent,
-    PoliciesChartComponent,
-  ],
-  exports: [PoliciesComponent, PolicyDetailsComponent, PoliciesChartComponent],
+  declarations: [PoliciesComponent, PolicyDetailsComponent],
+  exports: [PoliciesComponent, PolicyDetailsComponent],
   providers: [PolicyService, ChangeRequestService],
-  entryComponents: [
-    PoliciesComponent,
-    PolicyDetailsComponent,
-    PoliciesChartComponent,
-  ],
 })
 export class PolicyModule {}
