@@ -1,16 +1,12 @@
 import { Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
-import {
-  B2BUser,
-  OCC_USER_ID_ANONYMOUS,
-  UserIdService,
-  UserService,
-} from '@spartacus/core';
+import { B2BUser, OCC_USER_ID_ANONYMOUS, UserIdService } from '@spartacus/core';
 import { Observable } from 'rxjs/internal/Observable';
 import { filter, map, switchMap, take } from 'rxjs/operators';
 import * as fromAction from '../store/actions';
 import { StateWithProductAssignment } from '../store/product-assignments-state';
 import * as fromSelector from '../store/selectors';
+import { UserAccountFacade } from '@spartacus/user/account/root';
 @Injectable({
   providedIn: 'root',
 })
@@ -18,7 +14,7 @@ export class ProductAssignmentService {
   constructor(
     protected store: Store<StateWithProductAssignment>,
     protected userIdService: UserIdService,
-    protected userService: UserService
+    protected userAccountFacade: UserAccountFacade
   ) {}
 
   user: string;
@@ -70,8 +66,8 @@ export class ProductAssignmentService {
   }
 
   isUserAdminOfUnit(unitId: string): Observable<boolean> {
-    return this.userService.get().pipe(
-      filter(user => user?.uid !== OCC_USER_ID_ANONYMOUS),
+    return this.userAccountFacade.get().pipe(
+      filter(user => !!user && user?.uid !== OCC_USER_ID_ANONYMOUS),
       map(user => {
         return (<B2BUser>user).orgUnit?.uid === unitId;
       })
